+++ .env
@@ -3,4 +3,4 @@ NEXT_TELEMETRY_DISABLE=1
 TOKEN="EXAMPLE"
 NEXT_PUBLIC_API_URL=http://localhost:3000/api
 # EXAMPLE="ON"
-DATABASE_URL="postgresql://user:password@localhost:5432/mydb?schema=public"
+DATABASE_URL="postgresql://user:test12345678@localhost:5432/main?schema=public"
