+++ src/app/recipes/page.tsx
@@ -1,17 +1,11 @@
 import RecipeCard from "@/components/recipes/RecipeCard"
 import type { Recipe } from "@prisma/client"
 import type { Metadata } from "next"
+import { fetchRecipes } from "@/helpers/api"
+
 export const metadata: Metadata = {
   title: "Книга рецептов",
   description: "Книга рецептов. Делитесь любимыми рецептами и открывайте для себя новые",
-}
-
-async function fetchRecipes() {
-  const res = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/recipes`, { cache: "no-store" })
-  if (!res.ok) {
-    throw new Error("404")
-  }
-  return await res.json()
 }
 
 export default async function RecipesPage() {
