+++ src/helpers/api.tsx
@@ -1,0 +1,9 @@
+async function fetchRecipes() {
+  const res = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/recipes`, { cache: "no-store" })
+  if (!res.ok) {
+    throw new Error("404")
+  }
+  return await res.json()
+}
+
+export { fetchRecipes }
